Signal Tables

The functions below operate on a signal table that implements the Abstract Signal Table Interface.

Note

SignalTables.jl exports all symbols of this table.
Modia.jl reexports all symbols and uses instantiatedModel as signalTable argument.

Signal table functions	Description

SignalTable	Returns a new SignalTable dictionary.

showInfo	Writes info about a signal table to the output stream.

[independentSignalNames](@ref	Returns the names of the independent signals.

signalNames	Returns a string vector of the signal names that are present in a signal table.

hasSignal	Returns true if a signal is present in a signal table.

getSignal	Returns signal from a signal table as Var or as Par.

getSignalInfo	Returns signal with :_typeof, :_size keys instead of :values/:value keys.

getValues	Returns the values of a Var signal from a signal table.

getValuesWithUnit	Returns the values of a Var signal from a signal table including its unit.

getValue	Returns the value of a Par signal from a signal table.

getValueWithUnit	Returns the value of a Par signal from a signal table including its unit.

getFlattenedSignal	Returns a copy of a signal where the values or the value are flattened and converted for use in plots or traditional tables.

getDefaultHeading	Returns the default heading for a plot.

SignalTables.SignalTable — Type

sigTable = SignalTable(args...)

Returns a new SignalTable dictionary.

Arguments args... are dictionary pairs where values must be Var(...) or Par(...). Example:

using SignalTables
using Unitful

t = 0.0:0.1:0.5
sigTable = SignalTable(
  "time"         => Var(values= t, unit="s", independent = true),
  "load.r"       => Var(values= [sin.(t) cos.(t) sin.(t)]),
  "motor.angle"  => Var(values= sin.(t), unit="rad"),
  "motor.w_ref"  => Var(values= cos.(t), unit="rad/s", info="Reference"),
  "motor.w_m"    => Var(values= Clocked(0.9*cos.(t),factor=2), unit="rad/s", info="Measured"),
  "motor.inertia"=> Par(value = 0.02, unit="kg*m/s^2"),
  "attributes"   => Par(info  = "This is a test signal table")
)

The first argument must define the independent signal, that is, Var(values=..., independent=true), ... and values must be an AbstractVector. Further added signals with a :values key, must have the same first dimension as the independent signal.

Most dictionary operations can be applied on sigTable, as well as all functions of Overview of Functions.

Examples

using SignalTables
using Unitful

t = 0.0:0.1:0.5
sigTable = SignalTable(
  "time"         => Var(values= t, unit="s", independent=true),
  "load.r"       => Var(values= [sin.(t) cos.(t) sin.(t)], unit="m"),
  "motor.angle"  => Var(values= sin.(t), unit="rad", state=true),
  "motor.w"      => Var(values= cos.(t), unit="rad/s", integral="motor.angle"),
  "motor.w_ref"  => Var(values= 0.9*[sin.(t) cos.(t)], unit = ["rad", "1/s"],
                                info="Reference angle and speed"),
  "wm"           => Var(alias = "motor.w"),
  "ref.clock"    => Var(values= [true, missing, missing, true, missing, missing],
                                 variability="clock"),
  "ref.trigger"  => Var(values= [missing, missing, true, missing, true, true],
                                 variability="trigger"),
  "motor.w_c"    => Var(values= [0.8, missing, missing, 1.5, missing, missing],
                                variability="clocked", clock="ref.clock"),
  "motor.inertia"=> Par(value = 0.02f0, unit="kg*m/s^2"),
  "motor.data"   => Par(value = "resources/motorMap.json"),
  "attributes"   => Par(info  = "This is a test signal table")
)

signalInfo(sigTable)

This results in the following output:

name          unit          size  basetype kind attributes
─────────────────────────────────────────────────────────────────────────────────────────
time          "s"           (6,)  Float64  Var  independent=true
load.r        "m"           (6,3) Float64  Var
motor.angle   "rad"         (6,)  Float64  Var  state=true
motor.w       "rad/s"       (6,)  Float64  Var  integral="motor.angle"
motor.w_ref   ["rad","1/s"] (6,2) Float64  Var  info="Reference angle and speed"
wm            "rad/s"       (6,)  Float64  Var  integral="motor.angle", alias="motor.w"
ref.clock                   (6,)  Bool     Var  variability="clock"
ref.trigger                 (6,)  Bool     Var  variability="trigger"
motor.w_c                   (6,)  Float64  Var  variability="clocked", clock="ref.clock"
motor.inertia "kg*m/s^2"    ()    Float32  Par
motor.data                        String   Par
attributes                                 Par  info="This is a test signal table"

The command show(IOContext(stdout, :compact => true), sigTable) results in the following output:

SignalTable(
  "time" => Var(values=0.0:0.1:0.5, unit="s", independent=true),
  "load.r" => Var(values=[0.0 1.0 0.0; 0.0998334 0.995004 0.0998334; 0.198669 0.980067 0.198669; 0.29552 0.955336 0.29552; 0.389418 0.921061 0.389418; 0.479426 0.877583 0.479426], unit="m"),
  "motor.angle" => Var(values=[0.0, 0.0998334, 0.198669, 0.29552, 0.389418, 0.479426], unit="rad", state=true),
  "motor.w" => Var(values=[1.0, 0.995004, 0.980067, 0.955336, 0.921061, 0.877583], unit="rad/s", integral="motor.angle"),
  "motor.w_ref" => Var(values=[0.0 0.9; 0.0898501 0.895504; 0.178802 0.88206; 0.265968 0.859803; 0.350477 0.828955; 0.431483 0.789824], unit=["rad", "1/s"], info="Reference angle and speed"),
  "wm" => Var(values=[1.0, 0.995004, 0.980067, 0.955336, 0.921061, 0.877583], unit="rad/s", integral="motor.angle", alias="motor.w"),
  "ref.clock" => Var(values=Union{Missing, Bool}[true, missing, missing, true, missing, missing], variability="clock"),
  "ref.trigger" => Var(values=Union{Missing, Bool}[missing, missing, true, missing, true, true], variability="trigger"),
  "motor.w_c" => Var(values=Union{Missing, Float64}[0.8, missing, missing, 1.5, missing, missing], variability="clocked", clock="ref.clock"),
  "motor.inertia" => Par(value=0.02, unit="kg*m/s^2"),
  "motor.data" => Par(value="resources/motorMap.json"),
  "attributes" => Par(info="This is a test signal table"),
  )

SignalTables.showInfo — Function

showInfo([io::IO=stdout,] signalTable;
         sorted=false, Var=true, Par=true, attributes=true)

Writes info about a signal table to the output stream. The keyword arguments define what information shall be printed or whether the names shall be sorted or presented in definition order.

Example

using SignalTables
using Unitful

t = 0.0:0.1:0.5
sigTable = SignalTable(
  "time"         => Var(values= t, unit="s", independent=true),
  "load.r"       => Var(values= [sin.(t) cos.(t) sin.(t)], unit="m"),
  "motor.angle"  => Var(values= sin.(t), unit="rad", state=true),
  "motor.w"      => Var(values= cos.(t), unit="rad/s", integral="motor.angle"),
  "motor.w_ref"  => Var(values= 0.9*[sin.(t) cos.(t)], unit = ["rad", "1/s"],
                                info="Reference angle and speed"),
  "wm"           => Var(alias = "motor.w"),
  "ref.clock"    => Var(values= [true, missing, missing, true, missing, missing],
                                 variability="clock"),
  "ref.trigger"  => Var(values= [missing, missing, true, missing, true, true],
                                 variability="trigger"),
  "motor.w_c"    => Var(values= [0.8, missing, missing, 1.5, missing, missing],
                                variability="clocked", clock="ref.clock"),
  "motor.inertia"=> Par(value = 0.02f0, unit="kg*m/s^2"),
  "motor.data"   => Par(value = "resources/motorMap.json"),
  "attributes"   => Par(info  = "This is a test signal table")
)

signalInfo(sigTable)

results in the following output

name          unit          size  basetype kind attributes
─────────────────────────────────────────────────────────────────────────────────────────
time          "s"           (6,)  Float64  Var  independent=true
load.r        "m"           (6,3) Float64  Var
motor.angle   "rad"         (6,)  Float64  Var  state=true
motor.w       "rad/s"       (6,)  Float64  Var  integral="motor.angle"
motor.w_ref   ["rad","1/s"] (6,2) Float64  Var  info="Reference angle and speed"
wm            "rad/s"       (6,)  Float64  Var  integral="motor.angle", alias="motor.w"
ref.clock                   (6,)  Bool     Var  variability="clock"
ref.trigger                 (6,)  Bool     Var  variability="trigger"
motor.w_c                   (6,)  Float64  Var  variability="clocked", clock="ref.clock"
motor.inertia "kg*m/s^2"    ()    Float32  Par
motor.data                  ()    String   Par
attributes                                 Par  info="This is a test signal table"

SignalTables.independentSignalNames — Function

independentSignalNames(signalTable)::Vector{String}

Returns the names of the independent signals (often: ["time"]) from signalTable.

SignalTables.signalNames — Function

signalNames(signalTable)::Vector{String}

Returns a string vector of the signal names that are present in signalTable (including independent signal names).

SignalTables.hasSignal — Function

hasSignal(signalTable, name::String)

Returns true if signal name is present in signalTable.

SignalTables.getSignal — Function

getSignal(signalTable, name::String)

Returns signal name from signalTable (that is a Var or a Par). If name does not exist, an error is raised.

SignalTables.getSignalInfo — Function

getSignalInfo(signalTable, name::String)

Returns signal with :typeof, :size keys instead of :values/:value keys in form of a Var or a Par): where keys :values/:value are replaced by:

key	value

:_typeof	= typeof( signal[:values] ) or typeof( signal[:value] ) (if :value defined)

:_size	= size( signal[:values] ) or size( signal[:value] ) (if :value defined).

If name does not exist, an error is raised.

This function is useful if only the attributes of a signal are needed, but not their values (returning the attributes might be a cheap operation, whereas returning the values in the form required by the Abstract Signal Table Interface might be an expensive operation).

SignalTables.getValues — Function

getValues(signalTable, name)

Returns the values of a Var signal name from signalTable.

SignalTables.getValuesWithUnit — Function

getValuesWithUnit(signalTable, name)

Returns the values of a Var signal name from signalTable including its unit.

SignalTables.getValue — Function

getValue(signalTable, name)

Returns the value of a Par signal name from signalTable.

SignalTables.getValueWithUnit — Function

getValueWithUnit(signalTable, name)

Returns the value of a Par signal name from signalTable including its unit.

SignalTables.getFlattenedSignal — Function

flattenedSignal = getFlattenedSignal(signalTable, name;
                                     missingToNaN = true,
                                     targetInt    = Int,
                                     targetFloat  = Float64)

Returns a copy of a signal where the values or the value are flattened and converted (e.g.: missing -> NaN). A flattened signal can be, for example, used for traditional plot functions or for traditional tables.

Flattened values is a reshape of values into a vector or a matrix with optionally the following transformations:

name can be a signal name with or without array range indices (for example name = "a.b.c[2,3:5]").

If missingToNaN=true, then missing values are replaced by NaN values. If NaN does not exist in the corresponding type, the values are first converted to targetFloat.

If targetInt is not nothing, Int-types are converted to targetInt

If targetFloat is not nothing, Float-types are converted to targetFloat

collect(..) is performed on the result.

Legend is a vector of strings that provides a description for every array column (e.g. if "name=a.b.c[2,3:5]", unit="m/s", then legend = ["a.b.c[2,3] [m/s]", "a.b.c[2,3] [m/s]", "a.b.c[2,5] [m/s]"].

If the required transformation is not possible, a warning message is printed and nothing is returned.

SignalTables.getDefaultHeading — Function

getDefaultHeading(signalTable, name::String)::String

Returns the default heading for a plot.